update improvement for simulation handling
