--- src/main/java/com/pipiobjo/brewery/adapters/alarm/AlarmMockDevice.java
@@ -1,18 +1,0 @@
-package com.pipiobjo.brewery.adapters.alarm;
-
-public class AlarmMockDevice implements AlarmAdapter{
-    @Override
-    public void turnAlarmOn() {
-
-    }
-
-    @Override
-    public void turnAlarmOff() {
-
-    }
-
-    @Override
-    public void turnAlarmOnForMilliseconds(long milliseconds) {
-
-    }
-}
