+++ src/main/java/com/pipiobjo/brewery/adapters/spiextensionboard/SPIExtensionBoardMockAdapter.java
@@ -1,15 +1,28 @@
 package com.pipiobjo.brewery.adapters.spiextensionboard;
 
 import com.pipiobjo.brewery.services.simulation.BreweryHardwareSimulation;
+import io.quarkus.arc.profile.IfBuildProfile;
+import io.quarkus.runtime.LaunchMode;
+import io.quarkus.runtime.configuration.ProfileManager;
 import lombok.extern.slf4j.Slf4j;
 
+import javax.annotation.PostConstruct;
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
 @Slf4j
+@ApplicationScoped
+@IfBuildProfile("mockDevices")
 public class SPIExtensionBoardMockAdapter implements SPIExtensionBoardAdapter {
 
-    BreweryHardwareSimulation breweryHardwareSimulation = null;
+    @Inject
+    BreweryHardwareSimulation breweryHardwareSimulation;
 
-    public SPIExtensionBoardMockAdapter(BreweryHardwareSimulation breweryHardwareSimulation){
-        this.breweryHardwareSimulation = breweryHardwareSimulation;
+    @PostConstruct
+    void init() {
+        String activeProfile = ProfileManager.getActiveProfile();
+        String launchMode = LaunchMode.current().name();
+        log.info("Selecting mocking device for spi extension board, profile={}, launchMode={}", activeProfile, launchMode);
     }
 
     @Override
